api/requirements.txt
fastapi
uvicorn
kafka-python
sqlalchemy
psycopg2-binary
pydantic